\name{as.kernelMatrix}
\docType{methods}
\alias{as.kernelMatrix}
\alias{as.kernelMatrix-methods}
\alias{as.kernelMatrix,matrix-method}
\title{Assing kernelMatrix class to matrix objects}

\description{\code{as.kernelMatrix} in package \pkg{in Package `kernlab'} can be used
to assing the kernelMatrix class to matrix objects representing a kernel matrix. This 
matrixes can then be used with the kernelMatrix interfaces which most of the functions 
in \pkg{'kernlab'} support.}

\usage{
\S4method{as.kernelMatrix}{matrix}(x)
}
\arguments{
\item{x}{matrix to be assinged the \code{kernelMatrix} class }
}

\author{
  Alexandros Karatzoglou \cr
    \email{alexandros.karatzoglou@ci.tuwien.ac.at}
    }

\seealso{\code{\link{kernelMatrix}}, \code{\link{dots}}}
    
\keyword{methods}


\examples{


}
