\name{reuters}
\alias{reuters}
\alias{rlabels}
\title{Reuters Text Data}
\description{A small sample from the Reuters news data set.}
\usage{data(reuters)}

\format{
  A list of 40 text documents along with the labels. \code{reuters}
  contains the text documents and \code{rlabels} the labels in a vector.

}
\details{
  This dataset contains a list of 40 text documents along with the
  labels. The data consist out of 20 documents from the \code{acq}
  category and 20 documents from the crude category. The labels are
  stored in \code{rlabels}
  

}
\source{Reuters}
\keyword{datasets}
