\name{spirals}
\alias{spirals}
\title{Spirals Dataset}
\description{A toy data set representing
two spirals with Gaussian noise. The data was created with 
the \code{mlbench.spirals} function in \code{mlbench}.
}
\usage{data(spirals)}
\format{
A matrix with 300 observations and 2 variables.
}

\examples{
data(spirals)
plot(spirals)
}
\keyword{datasets}
